requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-excel-response==2.0.5
django-import-export
django-filter==23.5
django-jazzmin==2.6.0
django-redis==5.4.0
django-silk==5.2.0
djangorestframework==3.14.0
pycodestyle==2.12.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.1.1
sqlparse==0.4.4
tablib==3.5.0  # Tablibning to'g'ri versiyasi
tomli==2.0.2
typing_extensions==4.8.0
whitenoise==6.7.0
psycopg2-binary==2.9.6  # PostgreSQL uchun drayver
gunicorn
python-dotenv
dj-database-url
